+++ js/widgets/card/cardrenderer.anarchy1.js
@@ -77,6 +77,7 @@ Object.assign(CardRenderer, {
     combatBody += this._weaponBlock(pnj, weapons, "anarchy1", deps);
     combatBody += this._spellsBlock(pnj, spells, "anarchy1");
     combatBody += this._complexFormsBlock(pnj, complexForms, "anarchy1");
+    combatBody += this._statusRow(pnj, "anarchy1", deps);
     combatBody += this._drugRow(pnj, "anarchy1", deps);
     combatBody += this._vehicleChipRow(pnj, deps);
     combatBody += this._spiritChipRow(pnj, deps);